requirements.txt
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-binance>=1.0.19
numpy>=1.24.0
ta>=0.10.0
aiohttp>=3.9.0
asyncio>=3.4.3
